fix cmdres

+++ trunk/master/database/datapump/run-postrefresh.sh
@@ -18,6 +18,12 @@ function f_execute_all() {
 	S_REMOTE_ROOT=$C_ENV_CONFIG_REMOTE_ROOT
 	S_LOGDIR=$P_LOGDIR
 
+	f_execute_cmdres $S_REMOTE_HOSTLOGIN $S_REMOTE_ROOT "if [ -d $S_REMOTE_ROOT ]; then date > laststartdate.txt; echo ok; fi"
+	if [ "$S_RUNCMDRES" != "ok" ]; then
+		echo unable to access remote root - $S_REMOTE_ROOT. Exiting
+		exit 1
+	fi
+
 	# cleanup
 	rm -rf ./$S_LOGDIR
 	mkdir -p ./$S_LOGDIR
@@ -26,7 +32,7 @@ function f_execute_all() {
 	rm -rf ./$F_SQLDIR
 
 	F_SQLDIR_REMOTE=postrefresh
-	f_execute_cmd "rm -rf $F_SQLDIR_REMOTE"
+	f_execute_cmd $S_REMOTE_HOSTLOGIN $S_REMOTE_ROOT "rm -rf $F_SQLDIR_REMOTE"
 
 	# copy scripts and helper
 	F_SVNPATH=$C_CONFIG_SVNPATH/releases/$C_CONFIG_PRODUCT/database/refresh/$P_REFRESHDIR
@@ -36,11 +42,11 @@ function f_execute_all() {
 	scp datapump-config.sh $S_REMOTE_HOSTLOGIN:$S_REMOTE_ROOT
 	scp common.sh $S_REMOTE_HOSTLOGIN:$S_REMOTE_ROOT
 	scp import_helper.sh $S_REMOTE_HOSTLOGIN:$S_REMOTE_ROOT
-	f_execute_cmd "chmod 744 *.sh"
+	f_execute_cmd $S_REMOTE_HOSTLOGIN $S_REMOTE_ROOT "chmod 744 *.sh"
 
 	# execute remotedly
 	echo execute post-refresh scripts ...
-	f_execute_cmd "./import_helper.sh $P_ENV $P_DB $P_DBCONN postrefresh"
+	f_execute_cmd $S_REMOTE_HOSTLOGIN $S_REMOTE_ROOT "./import_helper.sh $P_ENV $P_DB $P_DBCONN postrefresh"
 
 	# copy log files
 	echo download log files to $F_SQLDIR_REMOTE ...
